Handle ReplicaDiskSoftAntiAffinity during upgrade
